"cspp2_m10_a1_1: 3/5 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ cspp2-assignments/Wallet/User.java
@@ -64,7 +64,7 @@ public class User {
 	public void balance(String name) {
 		for (int i = 0; i < size; i++) {
 			if (wallet[i].getName().equals(name)) {
-				System.out.println(wallet[i].getAmount());
+				System.out.println((float)wallet[i].getAmount());
 				break;
 			}
 		}
